requirements.txt
azure-servicebus==7.12.3
azure-storage-blob==12.19.1
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
ultralytics==8.2.103
Pillow==10.4.0
orjson==3.10.7
